FavoriteService created and injection functioning
